Reorganization of JAXB code for pagecontent
modified beans are now in package pagecontent_trp
TrpRegionType was introduced as middle class between RegionType and all
descendants of RegionType and is used for all the customizations of
RegionType that were formerly in RegionType itself. This is done by
using xjc with the -b flag and binding xml pagecontent_jaxb_binding.xml.
The script for code generation is in src/main/resources/scripts